Implement rule book classes
The specifications are fixed by introducing an interface for each rule book. In addition, fix a
specification for the category rule book.
