+++ src/main/kotlin/ch/schulealtendorf/sporttagpsa/business/rulebook/ResultRuleBook.kt
@@ -36,39 +36,23 @@
 
 package ch.schulealtendorf.sporttagpsa.business.rulebook
 
-import ch.schulealtendorf.rules.BasicRuleBook
-import org.springframework.stereotype.Component
-
 /**
- * Contains all the rules that are used in a PSA sport tournament.
+ * Describes a rule book for result calculation.
  * 
  * @author nmaerchy
- * @version 1.1.0
+ * @version 1.0.0
  */
-@Component
-class PSARuleBook: BasicRuleBook<FormulaModel, Int>(
-        FormulaModel::class,
-        Int::class
-), ResultRuleBook {
-    
-    init {
-        addRuleSet(SprintRuleSet())
-        addRuleSet(SkippingRuleSet())
-        addRuleSet(TargetThrowingRuleSet())
-        addRuleSet(BroadJumpRuleSet())
-        addRuleSet(BallThrowingRuleSet())
-        addRuleSet(BasketThrowingRuleSet())
-    }
+interface ResultRuleBook {
 
     /**
      * Runs a fact in this rulebook and calculates
      * the result depending on it.
-     *
+     * 
      * The result is at least 1.
      * 
      * @param fact the fact to run
-     *
+     * 
      * @return the resulting result
      */
-    override fun calc(fact: FormulaModel): Int = run(fact) ?: 1
+    fun calc(fact: FormulaModel): Int
 }
